requirements.txt
llama-cpp-python
mesa
networkx
chromadb
faiss-cpu
fastapi
uvicorn
streamlit
